Consolidate Operations DAO.
